api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
bcrypt
python-multipart
pillow
pydantic-settings
